requirements.txt
# Django Core
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0

# Database
psycopg2-binary>=2.9.0  # PostgreSQL adapter

# Task Queue & Caching
celery>=5.3.0
redis>=5.0.0
django-celery-beat>=2.5.0  # For periodic tasks

# Authentication & Security
django-allauth>=0.52.0  # Optional: for social auth
cryptography>=41.0.0

# Payment Processing
stripe>=7.0.0
requests>=2.31.0  # For PayPal API (also used for other HTTP requests)

# Email & SMS
django-sendgrid-v5>=5.4.0  # Or use django-mailgun
twilio>=8.10.0

# Calendar Integrations
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
msal>=1.24.0  # Microsoft Authentication Library
requests>=2.31.0  # For CalDAV/HTTP requests
icalendar>=5.0.0  # For iCal file generation

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
icalendar>=5.0.0  # For iCal file generation
django-timezone-field>=5.0
qrcode>=7.4.2  # For QR code generation
Pillow>=10.0.0  # Required by qrcode

# API Documentation
drf-yasg>=1.21.7  # Swagger/OpenAPI docs

# Environment Variables
python-decouple>=3.8

# Image Processing (for staff photos)
Pillow>=10.0.0

# CSV Export
django-import-export>=3.3.0

# Testing
pytest>=7.4.0
pytest-django>=4.7.0
pytest-cov>=4.1.0
factory-boy>=3.3.0

# Development
ipython>=8.15.0
django-debug-toolbar>=4.2.0
django-extensions>=3.2.0

